requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
google-cloud-firestore
numpy
python-dotenv